# PERF: keep using ObjectEngine for ExtensionArrays for 1.5